added testing for batch scraping
